jenkinsfile: add condition for cleanup

+++ Jenkinsfile
@@ -8,6 +8,9 @@ void setBuildStatus(String message, String state) {
   ]);
 }
 
+def  ubuntu18image = false
+def  ubuntu20image = false
+def  debianbusterimage = false
 
 pipeline {
      agent any
@@ -83,6 +86,9 @@ pipeline {
                             BRANCH_NAME == 'master' || BRANCH_NAME == 'development'
                        }
                     }
+                    script {
+                      debianbusterimage = true
+                    }
                     steps {
                      sh "docker build -f aecid-testsuite/docker/Dockerfile_deb -t aecid/aminer-debian-buster:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID --build-arg=varbranch=development --build-arg=vardistri=debian:buster ."
                      sh "docker run --rm aecid/aminer-debian-buster:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID"
@@ -94,6 +100,9 @@ pipeline {
                        expression {
                             BRANCH_NAME == 'master' || BRANCH_NAME == 'development'
                        }
+                    }
+                    script {
+                      ubuntu18image = true
                     }
                     steps {
                      sh "docker build -f aecid-testsuite/docker/Dockerfile_deb -t aecid/aminer-ubuntu-1804:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID --build-arg=varbranch=development --build-arg=vardistri=ubuntu:18.04 ."
@@ -107,6 +116,9 @@ pipeline {
                             BRANCH_NAME == 'master' || BRANCH_NAME == 'development'
                        }
                     }
+                    script {
+                      ubuntu20image = true
+                    }
                     steps {
                      sh "docker build -f aecid-testsuite/docker/Dockerfile_deb -t aecid/aminer-ubuntu-2004:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID --build-arg=varbranch=development --build-arg=vardistri=ubuntu:20.04 ."
                      sh "docker run --rm aecid/aminer-ubuntu-1804:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID"
@@ -116,9 +128,15 @@ pipeline {
     post {
         always {
            sh "docker rmi aecid/logdata-anomaly-miner-testing:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID"
-           sh "docker rmi aecid/aminer-debian-buster:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID"
-           sh "docker rmi aecid/aminer-ubuntu-1804:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID"
-           sh "docker rmi aecid/aminer-ubuntu-2004:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID"
+           if( debianbusterimage ){
+               sh "docker rmi aecid/aminer-debian-buster:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID"
+           }
+           if( ubuntu18image ){
+               sh "docker rmi aecid/aminer-ubuntu-1804:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID"
+           }
+           if( ubuntu20image ){
+               sh "docker rmi aecid/aminer-ubuntu-2004:$JOB_BASE_NAME-$EXECUTOR_NUMBER-$BUILD_ID"
+           }
         }
 	success {
         setBuildStatus("Build succeeded", "SUCCESS");
